Creating DeeplinkExtension to map into screenstack
